+++ app/src/main/java/fr/bendev/seatgeekapp/pages/event/EventDetailUIState.kt
@@ -1,0 +1,7 @@
+package fr.bendev.seatgeekapp.pages.event
+
+import fr.bendev.seatgeekapp.domain.model.Event
+
+data class EventDetailUIState(
+    val event: Event? = null
+)