+++ framework/src/main/java/fr/bendev/seatgeekapp/framework/network/model/ApiEvent.kt
@@ -15,5 +15,7 @@ data class ApiEvent(
     @SerializedName("venue")
     val venue: ApiVenue,
     @SerializedName("stats")
-    val stats: ApiStats
+    val stats: ApiStats,
+    @SerializedName("performers")
+    val performers: List<ApiPerformer>,
 )