+++ framework/src/main/java/fr/bendev/seatgeekapp/framework/network/model/ApiPerformer.kt
@@ -1,0 +1,8 @@
+package fr.bendev.seatgeekapp.framework.network.model
+
+import com.google.gson.annotations.SerializedName
+
+data class ApiPerformer(
+    @SerializedName("image")
+    val image: String,
+)