+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.0.1
 * Init project
 * Create list of events from API
+* Create page for event details